styles: adjust style landing page
